pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pydantic = "^2.8.2"
playwright = "^1.46.0"
python-dotenv = "^1.0.1"
finicapi = "^0.1.27"

[tool.poetry.scripts]
start = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
